Pipe incarnations through C plugin.
Recall that when a task in a multi-controller JAX (mcjax) job starts, it is
given a randomly generated incarnation id. If the task fails and restarts, it
gets a new id. These incarnation ids help us distinguish two different
instances of the same task.

When a group of tasks wants to perform a collective (e.g., AllReduce), every
participate needs to agree on the set of participants. If task 0 thinks the
collective is with tasks {0, 1, 2} while task 1 thinks the collective is with
tasks {0, 1, 3}, the program will not execute smoothly.

The participants also need to agree on everyone's incarnations. For example, if
task 0 thinks task 2 has incarnation 42 while task 1 thinks task 2 has
incarnation 100, they will disagree on whether task 2 has failed or not.

Tasks get the set of participants and their incarnations from calls to
`live_devices`. `live_devices` runs a barrier across tasks and returns the set
of currently alive processes along with their incarnations. This CL pipes this
information through the C API, so that open source clients can use the
information. Concretely, the incarnations of every task are passed to an
executable when it is run.

PiperOrigin-RevId: 786865712